Extend tests to support daemon
Add features to tests to account for the fact that
daemon is doing push or fetch in the background.

The user's original push should now finish more
quickly, but we now can't depend on the other repos
having the same ref as soon as it's done.

Daemon still off until all tests pass.